requirements.txt
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
isodate==0.6.1
lxml==4.9.2
platformdirs==2.6.2
PyJWT==2.6.0
python-dotenv==0.21.0
pytz==2022.7.1
requests==2.28.2
requests-file==1.5.1
requests-toolbelt==0.10.1
simple-salesforce==1.12.3
six==1.16.0
urllib3==1.26.14
zeep==4.2.1
